requirements.txt
requests==2.32.3
pandas==2.2.3
python-dotenv==1.0.0
mysql-connector==2.2.9
mysql--connector-python ==9.1.0
SQLAlchemy==2.0.7
# apache-airflow==2.10.4
# apache-airflow-providers-google==12.0.0
# google-cloud-bigquery==3.11.4